Fix regex for case insensitive headers
I believe that upgrading to protocol HTTP 2.0 causes the header names to become lowercase:

Just as in HTTP/1.x, header field names are strings of ASCII characters that are compared in a case-insensitive fashion. However, header field names MUST be converted to lowercase prior to their encoding in HTTP/2. A request or response containing uppercase header field names MUST be treated as malformed
-- RFC 7540, Hypertext Transfer Protocol Version 2 (HTTP/2), Section 8.1.2